Cleanup install instructions for optional libs (#2139)
This cleans up the install instructions to specify that optional libs
should be installed using the `kafka-python[opt]` format. This leverages
`setuptools`' `extra_requires` feature to let our users choose to
inherit any versions pins we might apply to our optional dependencies.

This also re-orders the docs slightly to give more visibility to the
`crc32c` install because users are unlikely to realize it exists.

It also cleans up the formatting of the compression libraries slightly,
as they were getting a little unwieldy.